refactor(timelock): centralize BIP68 encode/decode helpers
