+++ src/contracts/handlers/helpers.ts
@@ -1,5 +1,8 @@
 import { RelativeTimelock } from "../../script/tapscript";
-import * as bip68 from "bip68";
+import {
+    timelockToSequence as timelockToSequenceImpl,
+    sequenceToTimelock as sequenceToTimelockImpl,
+} from "../../utils/timelock";
 import { Contract, PathContext } from "../types";
 import { isDescriptor, extractPubKey } from "../../identity/descriptor";
 
@@ -23,25 +26,14 @@ function extractRawPubKey(value: string): string | undefined {
  * Convert RelativeTimelock to BIP68 sequence number.
  */
 export function timelockToSequence(timelock: RelativeTimelock): number {
-    return bip68.encode(
-        timelock.type === "blocks"
-            ? { blocks: Number(timelock.value) }
-            : { seconds: Number(timelock.value) }
-    );
+    return timelockToSequenceImpl(timelock);
 }
 
 /**
  * Convert BIP68 sequence number back to RelativeTimelock.
  */
 export function sequenceToTimelock(sequence: number): RelativeTimelock {
-    const decoded = bip68.decode(sequence);
-    if ("blocks" in decoded && decoded.blocks !== undefined) {
-        return { type: "blocks", value: BigInt(decoded.blocks) };
-    }
-    if ("seconds" in decoded && decoded.seconds !== undefined) {
-        return { type: "seconds", value: BigInt(decoded.seconds) };
-    }
-    throw new Error(`Invalid BIP68 sequence: ${sequence}`);
+    return sequenceToTimelockImpl(sequence);
 }
 
 /**
